Implements `identifier`
In 10.10, something in the class hierarchy of MPPaletteViewController implements NSUserInterfaceItemIdentification, which adds an `identifier` string property to the class. 10.9 and earlier doesn't do that, so instead we crash when trying to set the identifier.

This adds the identifier property, and uses the system implementation of the method if it exists.